패스워드 공통 처리. 테스트 필

+++ src/main/java/com/utime/burrowNest/admin/vo/ManageUserVo.java
@@ -15,9 +15,9 @@ public class ManageUserVo extends UserVo {
 	/**
 	 * 총 로그인 횟수
 	 */
-	int loginCount;
+	long loginCount;
 	/** 마지막 로그인 날짜 */
 	LocalDateTime lastLoginTime;
 	/** 로그인 실패 횟수 */
-	int loginFailCount;
+	long loginFailCount;
 }
